+++ gym_super_mario_bros/_roms/__init__.py
@@ -1,10 +1,12 @@
 """Methods for ROM file management."""
 from .decode_target import decode_target
 from .rom_path import rom_path
+from .rom_path import smb2_rom_path
 
 
 # explicitly define the outward facing API of this package
 __all__ = [
     decode_target.__name__,
     rom_path.__name__,
+    smb2_rom_path.__name__,
 ]